Refactored host code to have a minimal oe verification library.
